requirements.txt
Django>=2.2
Markdown
celery==4.3.0
cx_Oracle==7.3
dj-database-url
django-colorfield
django-cors-headers
django-filter
django-grappelli
django-nested-admin
django-ordered-model==3.2.0
django-redis-cache==2.1.0
djangorestframework>=3.9
git+https://github.com/MinisterioPublicoRJ/login-sca
gunicorn==19.9.0
happybase==1.2.0
impyla
newrelic==5.4.0.132
psycopg2-binary
pyjwt
python-decouple
redis
requests
responses
six==1.14.0
zeep==3.4.0
Unipath==1.1
